move inertial_length tests into tests_lengths.py
